Refactor fec module into submodules
